Move source code to src/ directory
- Create src/ directory and move all source files except main.js
- Update import paths in index.html and test files
- Update Vite config aliases to reflect new paths
- Keep index.html in project root as per convention
