+++ index.html
@@ -28,6 +28,6 @@
       </div>
       <div class="footer"></div>
     </div>
-    <script type="module" src="/entry.js"></script>
+    <script type="module" src="/src/entry.js"></script>
   </body>
 </html>
